* added mkdir /report

+++ postgresql_benchmark/psb_run.py
@@ -6,11 +6,12 @@
 # ;===========================================================
 
 import argparse
+import os
 import subprocess
 
 from sys import exit
-from os import getuid
-from psb_conf import SCRIPT_DIR, LOG_FILENAME, REPORT_FILENAME, \
+from os import getuid, path
+from psb_conf import SCRIPT_DIR, LOG_FILENAME, REPORT_FILENAME, REPORT_PATH, \
     MAC_SQL_UPGRADE, MAC_SQL_TRANSACTION, \
     MIC_SQL_UPGRADE, MIC_SQL_TRANSACTION, \
     ACL_SQL_UPGRADE, ACL_SQL_TRANSACTION, \
@@ -59,8 +60,14 @@ args = parser.parse_args()
 if getuid() != 0:
     exit(2)
 
-report = open(LOG_FILENAME, 'w')
-report.close()
+# clean log
+if path.exists(LOG_FILENAME):
+    report = open(LOG_FILENAME, 'w')
+    report.close()
+
+# create dir
+if not path.exists(REPORT_PATH):
+    os.mkdir(REPORT_PATH, mode=0o755)
 
 version = astra_version()
 if args.DB_PREPARE:
